improve code quality Divided duke package into smaller subpackages.
Divide parse method into smaller methods to avoid long deep nested code.
