+++ src/main/java/duke/Launcher.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.gui.Main;
 import javafx.application.Application;
 
 public class Launcher {
