+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,4 +1,7 @@
-package duke;
+package duke.command;
+
+import duke.processor.Executable;
+import duke.processor.TaskList;
 
 /**
  * Command that deletes a task from the list by its index.
