+++ src/main/java/duke/gui/MainWindow.java
@@ -1,5 +1,8 @@
-package duke;
+package duke.gui;
 
+import duke.processor.Duke;
+import duke.exception.DukeExcpetion;
+import duke.processor.Parser;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
