+++ src/main/java/duke/processor/Executable.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.processor;
+
+import duke.processor.TaskList;
 
 /**
  * type to return after parsing
