+++ src/main/java/duke/processor/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.processor;
 
 import java.time.format.DateTimeFormatter;
