+++ src/main/java/duke/task/Deadline.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.task;
+
+import duke.processor.Ui;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
